tests/wpt/tests/css/css-inline/text-box-trim/inheritance.html

<!DOCTYPE html>
<title>Inheritance of CSS Inline Layout properties</title>
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#propdef-text-box-edge">
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#propdef-text-box-trim">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
<div id="container">
  <div id="target"></div>
</div>
<script>
assert_inherited('text-box-edge', 'leading', 'text');
assert_not_inherited('text-box-trim', 'none', 'start');
</script>
