LayoutTests/imported/w3c/web-platform-tests/css/css-inline/text-box-trim/inheritance-expected.txt


FAIL Property text-box-edge has initial value leading assert_equals: expected "leading" but got "leading leading"
FAIL Property text-box-edge inherits assert_equals: expected "text" but got "text text"
PASS Property text-box-trim has initial value none
PASS Property text-box-trim does not inherit
